+++ lib/omniauth-rescuetime.rb
@@ -1,2 +1,0 @@
-require "omniauth-rescuetime/version"
-require 'omniauth/strategies/rescuetime'