+++ lib/omniauth-sharefile.rb
@@ -1,0 +1,2 @@
+require "omniauth-sharefile/version"
+require 'omniauth/strategies/sharefile'