+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
-## omniauth-rescuetime - RELEASE NOTES
+## omniauth-sharefile - RELEASE NOTES
 
 ####Version 0.0.1 (July 29, 2014)
 
-* The first version of the RescueTime omniauth strategy
+* The first version of the Sharefile omniauth strategy